+++ src/CSharpLanguageServer/Handlers/CallHierarchy.fs
@@ -32,17 +32,20 @@ module CallHierarchy =
         |> Option.bind (fun x -> x.DynamicRegistration)
         |> Option.defaultValue false
 
-    let provider (clientCapabilities: ClientCapabilities option) : bool option =
-        match dynamicRegistration clientCapabilities with
+    let provider (clientCapabilities: ClientCapabilities) : U3<bool, CallHierarchyOptions, CallHierarchyRegistrationOptions> option =
+        match dynamicRegistration (Some clientCapabilities) with
         | true -> None
-        | false -> Some true
+        | false -> Some (U3.C1 true)
 
     let registration (clientCapabilities: ClientCapabilities option) : Registration option =
         match dynamicRegistration clientCapabilities with
         | false -> None
         | true ->
             let registerOptions: CallHierarchyRegistrationOptions =
-                { DocumentSelector = Some defaultDocumentSelector }
+                { DocumentSelector = Some defaultDocumentSelector
+                  Id = None
+                  WorkDoneProgress = None
+                }
             Some
                 { Id = Guid.NewGuid().ToString()
                   Method = "textDocument/prepareCallHierarchy"
@@ -51,7 +54,7 @@ module CallHierarchy =
     let prepare (context: ServerRequestContext) (p: CallHierarchyPrepareParams) : AsyncLspResult<CallHierarchyItem[] option> = async {
         match! context.FindSymbol p.TextDocument.Uri p.Position with
         | Some symbol when isCallableSymbol symbol ->
-            let! itemList = HierarchyItem.fromSymbol context.ResolveSymbolLocations symbol
+            let! itemList = CallHierarchyItem.fromSymbol context.ResolveSymbolLocations symbol
             return
                 itemList
                 |> List.toArray
@@ -71,7 +74,7 @@ module CallHierarchy =
                 |> Seq.toArray
             info.CallingSymbol.Locations
             |> Seq.map (fun loc ->
-                { From = HierarchyItem.fromSymbolAndLocation (info.CallingSymbol) (loc |> Location.fromRoslynLocation)
+                { From = CallHierarchyItem.fromSymbolAndLocation (info.CallingSymbol) (loc |> Location.fromRoslynLocation)
                   FromRanges = fromRanges })
 
         match! context.FindSymbol p.Item.Uri p.Item.Range.Start with
@@ -91,7 +94,7 @@ module CallHierarchy =
     }
 
     let outgoingCalls
-        (context: ServerRequestContext)
+        (_context: ServerRequestContext)
         (_: CallHierarchyOutgoingCallsParams)
         : AsyncLspResult<CallHierarchyOutgoingCall[] option> = async {
         // TODO: There is no memthod of SymbolFinder which can find all outgoing calls of a specific symbol.