+++ src/CSharpLanguageServer/Handlers/Definition.fs
@@ -18,23 +18,24 @@ module Definition =
         |> Option.bind (fun x -> x.DynamicRegistration)
         |> Option.defaultValue false
 
-    let provider (clientCapabilities: ClientCapabilities option) : bool option =
-        match dynamicRegistration clientCapabilities with
+    let provider (clientCapabilities: ClientCapabilities) : U2<bool, DefinitionOptions> option =
+        match dynamicRegistration (Some clientCapabilities) with
         | true -> None
-        | false -> Some true
+        | false -> Some (U2.C1 true)
 
     let registration (clientCapabilities: ClientCapabilities option) : Registration option =
         match dynamicRegistration clientCapabilities with
         | false -> None
         | true ->
             let registerOptions: DefinitionRegistrationOptions =
-                { DocumentSelector = Some defaultDocumentSelector }
+                { DocumentSelector = Some defaultDocumentSelector
+                  WorkDoneProgress = None }
             Some
                 { Id = Guid.NewGuid().ToString()
                   Method = "textDocument/definition"
                   RegisterOptions = registerOptions |> serialize |> Some }
 
-    let handle (context: ServerRequestContext) (p: TextDocumentPositionParams) : AsyncLspResult<GotoResult option> = async {
+    let handle (context: ServerRequestContext) (p: TextDocumentPositionParams) : AsyncLspResult<Declaration option> = async {
         match! context.FindSymbol' p.TextDocument.Uri p.Position with
         | None -> return None |> success
         | Some (symbol, doc) ->
@@ -42,7 +43,7 @@ module Definition =
             return
                 locations
                 |> Array.ofList
-                |> GotoResult.Multiple
+                |> Declaration.C2
                 |> Some
                 |> success
     }