+++ src/CSharpLanguageServer/Types.fs
@@ -34,67 +34,13 @@ type ICSharpLspClient =
     // But ClientCapabilities is a complex type, write it again will be a huge work.
     abstract member Capabilities: ClientCapabilities option with get, set
 
-/// TODO: to be moved to Ionide.LanguageServerProtocol
-type TextDocumentRegistrationOptions =
-    { DocumentSelector: DocumentSelector option }
-
-/// TODO: to be moved to Ionide.LanguageServerProtocol
-type TextDocumentChangeRegistrationOptions =
-    { DocumentSelector: DocumentSelector option; SyncKind: TextDocumentSyncKind }
-
-/// TODO: to be moved to Ionide.LanguageServerProtocol
-type TextDocumentSaveRegistrationOptions =
-    { DocumentSelector: DocumentSelector option; IncludeText: bool option }
-
-/// TODO: to be moved to Ionide.LanguageServerProtocol
-type CallHierarchyRegistrationOptions =
-    { DocumentSelector: DocumentSelector option }
-
-/// TODO: to be moved to Ionide.LanguageServerProtocol
-type ReferenceRegistrationOptions =
-    { DocumentSelector: DocumentSelector option }
-
-/// TODO: to be moved to Ionide.LanguageServerProtocol
-type TypeDefinitionRegistrationOptions =
-    { DocumentSelector: DocumentSelector option }
-
-/// TODO: to be moved to Ionide.LanguageServerProtocol
-type DocumentFormattingRegistrationOptions =
-    { DocumentSelector: DocumentSelector option }
-
-/// TODO: to be moved to Ionide.LanguageServerProtocol
-type TypeHierarchyRegistrationOptions =
-    { DocumentSelector: DocumentSelector option }
-
-/// TODO: to be moved to Ionide.LanguageServerProtocol
-type DefinitionRegistrationOptions =
-    { DocumentSelector: DocumentSelector option }
-
-/// TODO: to be moved to Ionide.LanguageServerProtocol
-type ImplementationRegistrationOptions =
-    { DocumentSelector: DocumentSelector option }
-
-/// TODO: to be moved to Ionide.LanguageServerProtocol
-type DocumentRangeFormattingRegistrationOptions =
-    { DocumentSelector: DocumentSelector option }
-
-/// TODO: to be moved to Ionide.LanguageServerProtocol
-type DocumentHighlightRegistrationOptions =
-    { DocumentSelector: DocumentSelector option }
-
-/// TODO: to be moved to Ionide.LanguageServerProtocol
-type HoverRegistrationOptions =
-    { DocumentSelector: DocumentSelector option }
-
-/// TODO: to be moved to Ionide.LanguageServerProtocol
-type WorkspaceSymbolRegistrationOptions =
-    { ResolveProvider: bool option }
-
-type DocumentFilter with
-    static member Default: DocumentFilter =
+let defaultDocumentFilter: TextDocumentFilter =
         { Language = None
           Scheme = Some "file"
           Pattern = Some "**/*.cs" }
 
 // Type abbreviations cannot have augmentations, extensions
-let defaultDocumentSelector: DocumentSelector = [| DocumentFilter.Default |]
+let defaultDocumentSelector: DocumentSelector =
+    [|
+        defaultDocumentFilter |> U2.C1
+    |]